feature: meshsync as a library: move code from main to a library package
